backend/requirements.txt
fastapi
uvicorn
langchain
langchain-google-genai
langchain-anthropic
langchain-openai
python-dotenv 